+++ Mods ChoGGi/Remove Landscaping Limit/Code/Script.lua
@@ -1,7 +1,5 @@
 -- See LICENSE for terms
 
-local max_int = max_int
-local guim = guim
 local SetLandScapingLimits = ChoGGi_Funcs.Common.SetLandScapingLimits
 
 local mod_EnableMod